EEPROM bug squashed.
Well, this took a while to figure out. When you upload code using the
arduino bootloader, it appears that the EEPROM is wiped. This sets all
values in EEPROM to 255, including the byte that records the location
of the last address that has a card number :( After much fighting, it's
mostly fixed. EEPROM is still wiped when uploading new code, but the
system is now able to recover from that situation.